Move mixins under tools.mixins
