+++ python/marvin/tools/spaxel.py
@@ -8,31 +8,27 @@
 # @Copyright: José Sánchez-Gallego
 
 
-from __future__ import division
-from __future__ import print_function
-from __future__ import absolute_import
+from __future__ import absolute_import, division, print_function
 
 import abc
 import inspect
 import itertools
-import six
 import warnings
 
 import numpy as np
+import six
 
 import marvin
 import marvin.core.exceptions
-from marvin.core.exceptions import MarvinError, MarvinUserWarning, MarvinBreadCrumb
 import marvin.core.marvin_pickle
-import marvin.utils.general.general
-
 import marvin.tools.cube
 import marvin.tools.maps
 import marvin.tools.modelcube
-
-from marvin.utils.general.structs import FuzzyDict
+import marvin.utils.general.general
+from marvin.core.exceptions import MarvinBreadCrumb, MarvinError, MarvinUserWarning
 from marvin.utils.datamodel.dap import datamodel as dap_datamodel
 from marvin.utils.datamodel.drp import datamodel as drp_datamodel
+from marvin.utils.general.structs import FuzzyDict
 
 
 breadcrumb = MarvinBreadCrumb()
